<FIX> 쿠키 예외처리 추가
