+++ src/main/java/roomescape/GlobalExceptionController.java
@@ -1,0 +1,34 @@
+package roomescape;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.MissingRequestCookieException;
+import org.springframework.web.bind.annotation.ControllerAdvice;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import roomescape.exception.EntityNotFoundException;
+
+@ControllerAdvice
+public class GlobalExceptionController {
+    private static final Logger log = LoggerFactory.getLogger(GlobalExceptionController.class);
+
+    @ExceptionHandler(EntityNotFoundException.class)
+    public ResponseEntity<String> handleEntityNotFoundException(EntityNotFoundException e) {
+        log.info(e.getMessage());
+        e.printStackTrace();
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler(MissingRequestCookieException.class)
+    public ResponseEntity<String> handleMissingRequestCookieException(MissingRequestCookieException e) {
+        log.info(e.getMessage());
+        return ResponseEntity.badRequest().body(e.getMessage());
+    }
+
+    @ExceptionHandler(Exception.class)
+    public ResponseEntity<Void> handleRuntimeException(Exception e) {
+        e.printStackTrace();
+        return ResponseEntity.badRequest().build();
+    }
+
+}
